attributes to custom the owner/group
As the filename location can be changed with an attribute, it is
desirable to make the owner and group customizable, since someone may
wish to use a different user than root for AWS access.

For example, a sensu plugin may need to read the credentials when
running a plugin as the `sensu` user. A Jenkins server may need to
upload files to an S3 bucket as the `jenkins` user, etc.